pyproject.toml
[project]
name = "r1-vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "deepspeed>=0.16.3",
    "flash-attn>=2.7.4.post1",
    "imgcat>=0.6.0",
    "ipdb>=0.13.13",
    "jinja2>=3.1.5",
    "peft>=0.14.0",
    "pycocotools>=2.0.8",
    "python-dotenv>=1.0.1",
    "qwen-vl-utils>=0.0.10",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    # I need the latest version of transformers for qwen2.5vl, I should lock this down once they release a new version
    "transformers @ git+https://github.com/huggingface/transformers.git@main",
    "trl",
    "wandb>=0.19.5",
]

[tool.uv.sources]
trl = { path = "../trl", editable = true }